Use size_t instead of unsigned int to avoid compiler warning.
git-svn-id: http://scm.sipfoundry.org/rep/sipX/branches/sipXtapi@10674 a612230a-c5fa-0310-af8b-88eea846685b

+++ sipXportLib/src/os/OsFileSystem.cpp
@@ -313,7 +313,7 @@ OsStatus OsFileSystem::createDirRecursive(const OsPath& rOsPath)
         UtlString sep = OsPath::separator;
         
         int lastSep = -1;
-        unsigned int nextSep = rOsPath.index(sep);
+        size_t nextSep = rOsPath.index(sep);
         
         while (nextSep != UtlString::UTLSTRING_NOT_FOUND)
         {
